requirements.txt
pyopenssl
ndg-httpsclient
pyasn1
urllib3
request
scrapy
